Terceiro teste (testeMedia) falhando

+++ src/main/java/com/utfpr/classes/CalculaMediana.java
@@ -13,7 +13,7 @@ import java.util.Arrays;
  */
 public class CalculaMediana {
     int i;
-    float res = 0;
+    double res = 0;
     
     public double calculaMediana(double vetor[]){
         Arrays.sort(vetor);
@@ -23,13 +23,13 @@ public class CalculaMediana {
             if(vetor.length % 2 == 0){
                 int posicao1 = (int) ((vetor.length / 2) - 0.5f);
                 int posicao2 = (int) ((vetor.length / 2) + 0.5f);
-                res = (float) ((vetor[posicao1]+vetor[posicao2])/2);
+                res = (double) ((vetor[posicao1]+vetor[posicao2])/2);
                 return res;
             }else{
                 
                 if(vetor.length % 2 != 0){
                     int posicao = vetor.length / 2;
-                    res =(float) vetor[posicao];
+                    res =(double) vetor[posicao];
                     return res;
                 }
             }
